Move the add_numbers step
